Refactor KISSLoader to be more DRY.
